Save fetched data to local DB. Refresh when new value is fetched (#42)
With this cache rework, the page loads much faster, because old values
are stored in the cache. This may lead to potential issues if the newly
fetched value returns invalid value or fails. The old value maybe not
refreshed.

Widgets may use the first available value, which can be non recent. This
may result in strange results.

- Rework how fetching works to be fully sync.
- All cached calls require to provide `onInvalidate` callback.
- Move `selector` to async to speed up page load.