feat: update wallet-selector and add wallets (#252)
This PR updates the `Wallet Selector` to `v9.4.0`, which allows us now
to include Intear wallet. I further added `Hot Wallet` and updated
`Bitte` to its latest interface.

Interestingly, most of the changes on the PR actually deal with the fact
that `webpack` (or some other dependency) has clearly changed a lot on
the past 3 years. To manage to run the app locally I had to:

1. Add `vm-browserfy` as a dependency
2. Re-order the imports on app.jsx (otherwise a SES error is triggered)
3. Stop overwriting `props` on `Links` (otherwise a SES error is
triggered)
4. Add a config to stop `sass` from raising tons of `warnings`

---------

Co-authored-by: Guillermo Alejandro Gallardo Diez <gagdiez@MacBookPro.home>